# swyftr-swagger.yaml
swagger: '2.0'
info:
  version: '1'
  title: Swyftr
  description: Backend API for Swyftr system
host: 52.221.58.5
basePath: /api/v1
schemes:
  - https
...
paths:
...
  '/user/{userId}':
    get:
      summary: Get User Details
      tags:
        - User
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              userId:
                type: string
              firstName:
                type: string
              lastName:
                type: string
              dateOfBirth:
                type: string
              email:
                type: string
              phoneNo:
                type: number
              mobileNo:
                type: number
              nic:
                type: string
              addressList:
                type: array
                items:
                  $ref: '#/definitions/Address'
              profilePicUrl:
                type: string
        '400':
          description: Bad Request
    patch:
      summary: Update User
      tags:
        - User
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: userId
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              firstName:
                type: string
              lastName:
                type: string
              dateOfBirth:
                type: string
              profilePicId:
                type: number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
    delete:
      summary: Terminate User
      tags:
        - System
        - Sprint1
      description: Suspend a User
      produces:
        - application/json
      parameters:
        - in: path
          name: userId
          required: true
          type: string
      responses:
        '200':
          description: OK
        '500':
          description: Fail
  /user/address:
    post:
      summary: Add a new address to user
      tags:
        - User
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Address'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
    delete:
      summary: Remove address
      tags:
        - User
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: query
          name: addressId
          required: true
          type: integer
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
  /user/update/address:
    patch:
      summary: Update user address
      tags:
        - User
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: query
          name: addressId
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Address'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
  /user/cus/update_device_id:
    post:
      summary: Update device id of a user (push notifications)
      tags:
        - User
        - Sprint3
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              deviceId:
                type: string,
              deviceOs:
                type: string
                enum: ['ios', 'android']
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
  /user/cus/sendOTP:
    patch:
      summary: Re-send OTP
      tags:
        - User
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
        '400':
          description: Bad Request
        '500':
          description: Fail
  /user/cus/validateOTP:
    patch:
      summary: Validate OTP
      tags:
        - User
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              otp:
                type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
                type: string
        '400':
          description: Bad Request
        '500':
          description: Fail
...
definitions:
...
  Address:
    type: object
    properties:
      id:
        type: string
      userId:
        type: string
      addressLine1:
        type: string
      addressLine2:
        type: string
      district:
        type: string
      province:
        type: string
...